+++ package.json
@@ -53,9 +53,9 @@
     "lucide-react": "^0.462.0",
     "next-themes": "^0.3.0",
     "puppeteer": "^24.6.1",
-    "react": "^19.1.0",
+    "react": "^18.3.1",
     "react-day-picker": "^8.10.1",
-    "react-dom": "^19.1.0",
+    "react-dom": "^18.3.1",
     "react-helmet": "^6.1.0",
     "react-hook-form": "^7.53.0",
     "react-resizable-panels": "^2.1.3",
