+++ src/modules/products/components/AiGenerateButton.tsx
