+++ src/modules/products/components/List/ProductListRefactored.tsx
