+++ src/modules/seo/components/CompletionProgress.tsx
